requirements.txt
openai
langchain
streamlit==1.34.0
huggingface_hub
ollama
python-dotenv
# protobuf==3.19.0